Refactored search and added App to its own folder
